Async Get / Valor da taxa / Ajuste abstração / Redução de acomplamento com uso de interface

+++ 3 - Domain/Microservices.TaxaDeJuros.Domain/Builders/ITaxaDeJurosBuilder.cs
@@ -2,6 +2,6 @@
 {
     public interface ITaxaDeJurosBuilder<out T> : IBuilder<T>
     {
-        ITaxaDeJurosBuilder<T> WithValor(double valor);
+        ITaxaDeJurosBuilder<T> WithValor(decimal valor);
     }
 }
